Move RandomRune() pkg/rand/
This functionality will be needed outside of tests in the next commit so
move it out of testutils.
